Partially merge test_pil.py into test_io.py
